make standalone template cw-template-like structure
