--- framework/contracts/templates/standalone/src/handlers/mod.rs
@@ -1,6 +1,0 @@
-pub mod execute;
-pub mod instantiate;
-pub mod migrate;
-pub mod query;
-
-pub use self::{execute::execute, instantiate::instantiate, migrate::migrate, query::query};
