+++ framework/contracts/templates/standalone/src/lib.rs
@@ -1,18 +1,31 @@
 pub mod contract;
 pub mod error;
-pub mod handlers;
 pub mod msg;
-pub mod replies;
 pub mod state;
 
+use abstract_standalone::StandaloneContract;
+use cosmwasm_std::Response;
 pub use error::MyStandaloneError;
 
-pub(crate) use contract::MY_STANDALONE;
 /// The version of your standalone
 pub const STANDALONE_VERSION: &str = env!("CARGO_PKG_VERSION");
-
-pub use contract::interface::MyStandaloneInterface;
 
 pub const MY_NAMESPACE: &str = "yournamespace";
 pub const MY_STANDALONE_NAME: &str = "my-standalone";
 pub const MY_STANDALONE_ID: &str = const_format::formatcp!("{MY_NAMESPACE}:{MY_STANDALONE_NAME}");
+
+/// The type of the result returned by your standalone's entry points.
+pub type MyStandaloneResult<T = Response> = Result<T, MyStandaloneError>;
+
+/// The type of the standalone that is used to build your contract object and access the Abstract SDK features.
+pub type MyStandalone = StandaloneContract;
+
+pub const MY_STANDALONE: MyStandalone =
+    MyStandalone::new(MY_STANDALONE_ID, STANDALONE_VERSION, None);
+
+// cw-orch related interface
+#[cfg(not(target_arch = "wasm32"))]
+pub mod interface;
+
+#[cfg(not(target_arch = "wasm32"))]
+pub use interface::MyStandaloneInterface;
